requirements.txt
paho-mqtt
envirophat
